src/lasdamastdd/controllers/StartController.java
package lasdamastdd.controllers;

import lasdamastdd.models.Game;
import lasdamastdd.models.State;

public class StartController extends Controller {

    public StartController(Game game, State state) {
        super(game, state);
    }

    public void start() {
        this.game = new Game();
        this.state.next();
    }

    @Override
    public void accept(ControllersVisitor controllersVisitor) {
        assert controllersVisitor != null;
        controllersVisitor.visit(this);
    }
}
